Remove extra test results, move mobilenet to pretrained models
