+++ platform-tests/src/test/kotlin/org/nd4j/samediff/frameworkimport/onnx/loader/TestOnnxProcessLoader.kt
@@ -35,18 +35,7 @@ import org.nd4j.samediff.frameworkimport.registry.OpMappingRegistry
 
 class TestOnnxProcessLoader {
 
-    @Test
-    fun testMobileNet() {
-        System.setProperty("org.bytedeco.javacpp.nopointergc","true")
-        Nd4j.getExecutioner().enableDebugMode(true)
-        Nd4j.getExecutioner().enableVerboseMode(true)
-        val importer = OnnxFrameworkImporter()
-        val resource = ClassPathResource("mobilenet.onnx").file
-        val inputs = Nd4j.ones(1,3,224,224)
-        val inputDict = mapOf("input.1" to inputs)
-        val runImport = importer.runImport(resource.absolutePath,inputDict)
-        runImport.outputAll(inputDict)
-    }
+
 
     @Test
     fun testLoader() {
